# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Platform API",
        "description": "API complète pour la plateforme d'emploi - Authentification, gestion des offres et candidatures",
        "contact": {
            "email": "support@jobplatform.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Liste toutes les offres d'emploi actives",
                "description": "Retourne la liste de toutes les offres d'emploi actives avec les informations de l'employeur",
                "operationId": "9538d30d530d4c3680ea31fe228944e2",
                "responses": {
                    "200": {
                        "description": "Liste des offres d'emploi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Créer une nouvelle offre d'emploi",
                "description": "Crée une nouvelle offre d'emploi (réservé aux employeurs)",
                "operationId": "36deb2a6480382cf8385911c10aeebf0",
                "requestBody": {
                    "description": "Données de l'offre d'emploi",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Offre créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - rôle employeur requis",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized. Employer role required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Afficher une offre spécifique",
                "description": "Retourne les détails d'une offre d'emploi spécifique",
                "operationId": "6de31489b41fa51255446ac2ad15b7fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre d'emploi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de l'offre",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Offre non trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Mettre à jour une offre d'emploi",
                "description": "Met à jour une offre d'emploi existante (propriétaire ou admin seulement)",
                "operationId": "66762b778fbd20dd4fbd09628dabba17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre à modifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Données à mettre à jour",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JobUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Offre mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Supprimer une offre d'emploi",
                "description": "Supprime une offre d'emploi (propriétaire ou admin seulement)",
                "operationId": "5537fd998abc3a62eef8c09e4ae52000",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'offre à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Offre supprimée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/search": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Recherche d'offres d'emploi",
                "description": "Recherche des offres d'emploi par titre, entreprise ou localisation",
                "operationId": "0c0290a245c580348856b16cf6be0b81",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Terme de recherche dans le titre",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "company",
                        "in": "query",
                        "description": "Terme de recherche dans l'entreprise",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Terme de recherche dans la localisation",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de la recherche",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JobListResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@test.com"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T12:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "Job": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Développeur Full-Stack"
                    },
                    "description": {
                        "type": "string",
                        "example": "Nous recherchons un développeur full-stack expérimenté..."
                    },
                    "company": {
                        "type": "string",
                        "example": "Tech Solutions"
                    },
                    "location": {
                        "type": "string",
                        "example": "Paris"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 50000
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "employer": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "JobRequest": {
                "required": [
                    "title",
                    "description",
                    "company",
                    "location"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Développeur Full-Stack"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description du poste..."
                    },
                    "company": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Tech Solutions"
                    },
                    "location": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Paris"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 50000,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "JobUpdateRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Développeur Full-Stack Senior"
                    },
                    "description": {
                        "type": "string",
                        "example": "Nouvelle description..."
                    },
                    "company": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Tech Solutions SAS"
                    },
                    "location": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Lyon"
                    },
                    "salary": {
                        "type": "number",
                        "format": "float",
                        "example": 55000,
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "JobListResponse": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/Job"
                }
            },
            "DeleteResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Job deleted successfully"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}